Actualizacion varia por implementacion y ajustes en sponsorship

+++ EventosUy.Application/Interfaces/IRegisterTypeService.cs
@@ -7,7 +7,7 @@ namespace EventosUy.Application.Interfaces
 {
     public interface IRegisterTypeService
     {
-        public Task<Result<Guid>> CreateAsync(string name, string description, float price, int quota, Guid eventiId);
+        public Task<Result<Guid>> CreateAsync(string name, string description, decimal price, int quota, Guid eventiId);
         public Task<Result<DTRegisterType>> GetDTAsync(Guid id);
         public Task<Result<RegisterType>> GetByIdAsync(Guid id);
         public Task<Result<List<RegisterTypeCard>>> GetAllByEditionAsync(Guid editionId);
