+++ EventosUy.Domain/DTOs/DataTypes/DTRegisterType.cs
@@ -5,11 +5,11 @@
         public string Name { get; init; }
         public string Edition { get; init; }
         public string Description { get; init; }
-        public float Price { get; init; }
+        public decimal Price { get; init; }
         public int Quota { get; init; }
         public DateTimeOffset Created { get; init; }
 
-        public DTRegisterType(string name, string edition, string description, float price, int quota, DateTimeOffset created) 
+        public DTRegisterType(string name, string edition, string description, decimal price, int quota, DateTimeOffset created) 
         {
             Name = name;
             Edition = edition;
