+++ EventosUy.Domain/DTOs/DataTypes/DTSponsorship.cs
@@ -7,14 +7,14 @@ namespace EventosUy.Domain.DTOs.DataTypes
         public string Name { get; init; }
         public DateTimeOffset Created { get; init; }
         public DateOnly Expired { get; init; }
-        public float Amount { get; init; }
+        public decimal Amount { get; init; }
         public int Free { get; init; }
         public string Code { get; init; }
         public SponsorshipTier Tier { get; init; }
         public string Edition { get; init; }
         public string Institution { get; init; }
 
-        public DTSponsorship(string name, DateTimeOffset created, DateOnly expired, float amount, int free, string code, SponsorshipTier tier, string edition, string institution) 
+        public DTSponsorship(string name, DateTimeOffset created, DateOnly expired, decimal amount, int free, string code, SponsorshipTier tier, string edition, string institution) 
         {
             Name = name;
             Created = created;
