+++ EventosUy.Domain/Entities/RegisterType.cs
@@ -9,13 +9,13 @@ namespace EventosUy.Domain.Entities
         public Guid Id { get; init; }
         public string Name { get; init; }
         public string Description { get; private set; }
-        public float Price { get; init; }
+        public decimal Price { get; init; }
         public int Quota { get; init; }
         public bool Active { get; private set; }
         public DateTimeOffset Created { get; init; }
         public Guid Edition { get; init; }
 
-        private RegisterType(string name, string description, float price, int quota, Guid editionId) 
+        private RegisterType(string name, string description, decimal price, int quota, Guid editionId) 
         {
             Id = Guid.NewGuid();
             Name = name;
@@ -27,7 +27,7 @@ namespace EventosUy.Domain.Entities
             Created = DateTimeOffset.UtcNow;
         }
 
-        public static Result<RegisterType> Create(string name, string description, float price, int quota, Guid editionId) 
+        public static Result<RegisterType> Create(string name, string description, decimal price, int quota, Guid editionId) 
         {
             List<string> errors = [] ;
             if (string.IsNullOrWhiteSpace(name)) { errors.Add("Name cannot be empty."); }
